requirements.txt
aiosmtplib==2.0.2
alembic==1.13.1
anyio==4.3.0
bcrypt==4.0.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-mail==1.4.1
httpx==0.27.0
Jinja2==3.1.4
jose==1.0.0
psycopg2==2.9.9
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.30
starlette==0.37.2
uvicorn==0.29.0
SQLAlchemy-serializer==1.4.12